implement pathfinding

+++ include/Game/Components/CRangeMovement.h
@@ -2,20 +2,19 @@
 #define CRANGEMOVEMENT_H_20240128153242
 
 #include "Component.h"
-#include <cstddef>
 
 struct CRangeMovement
     : public snd::Component<CRangeMovement>
 {
 public:
-    size_t getMovementRange() { return movementRange_; };
-    void setMovementRange(size_t movementRange) { movementRange_ = movementRange; };
+    int getMovementRange() { return movementRange_; };
+    void setMovementRange(int movementRange) { movementRange_ = movementRange; };
 
-    CRangeMovement(size_t range)
+    CRangeMovement(int range)
         : movementRange_(range){};
 
 private:
-    size_t movementRange_;
+    int movementRange_;
 };
 
 #endif
